// .oldbk/frontend/src/api/types.ts
export type JsonPrimitive = string | number | boolean | null;
export type JsonValue = JsonPrimitive | JsonValue[] | { [key: string]: JsonValue };
// ...
export type DynamicFieldType =
  | "TEXT"
  | "TEXTAREA"
  | "NUMBER"
  | "MONEY"
  | "DATE"
  | "DATETIME"
  | "SELECT"
  | "MULTI_SELECT"
  | "SWITCH"
  | "MEMBER"
  | "DEPT"
  | "ATTACHMENT"
  | "IMAGE"
  | "AUTO_NO"
  | "RELATION"
  | "SUB_TABLE"
  | "ADDRESS"
  | "TAG"
  | "JSON";